- BrowseTracker: Remove warning about comparing signed and unsigned integers.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@13733 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/BrowseTracker/JumpTracker.cpp
@@ -849,7 +849,7 @@ void JumpTracker::JumpDataAdd(const wxString& inFilename, const long inPosn, con
     }
 
     // Make room for a new jump
-    while (m_ArrayOfJumpData.GetCount() >= maxJumpEntries)
+    while (m_ArrayOfJumpData.GetCount() >= (size_t)maxJumpEntries)
     {
         // Remove the oldest entry (index 0 is guaranteed to exist here)
         m_ArrayOfJumpData.RemoveAt(0);
